Fix read values type in ASN tests

+++ tests/foundation/test_asn1rd.c
@@ -123,7 +123,7 @@ test__asn1rd_read_int8__encoded_int8_0__returns_0(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_INT8_0);
 
-    int value = vscf_asn1rd_read_int8(asn1rd);
+    int8_t value = vscf_asn1rd_read_int8(asn1rd);
 
     TEST_ASSERT_EQUAL(0, value);
 
@@ -138,7 +138,7 @@ test__asn1rd_read_int8__encoded_int8_127__returns_127(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_INT8_127);
 
-    int value = vscf_asn1rd_read_int8(asn1rd);
+    int8_t value = vscf_asn1rd_read_int8(asn1rd);
 
     TEST_ASSERT_EQUAL(127, value);
 
@@ -153,7 +153,7 @@ test__asn1rd_read_int8__encoded_int8_neg_128__returns_neg_128(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_INT8_NEG_128);
 
-    int value = vscf_asn1rd_read_int8(asn1rd);
+    int8_t value = vscf_asn1rd_read_int8(asn1rd);
 
     TEST_ASSERT_EQUAL(-128, value);
 
@@ -168,7 +168,7 @@ test__asn1rd_read_int16__encoded_int16_32760__returns_32760(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_INT16_32760);
 
-    int value = vscf_asn1rd_read_int16(asn1rd);
+    int16_t value = vscf_asn1rd_read_int16(asn1rd);
 
     TEST_ASSERT_EQUAL(32760, value);
 
@@ -183,7 +183,7 @@ test__asn1rd_read_int16__encoded_int16_neg_32760__returns_neg_32760(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_INT16_NEG_32760);
 
-    int value = vscf_asn1rd_read_int16(asn1rd);
+    int16_t value = vscf_asn1rd_read_int16(asn1rd);
 
     TEST_ASSERT_EQUAL(-32760, value);
 
@@ -198,7 +198,7 @@ test__asn1rd_read_int32__encoded_int32_2147483000__returns_2147483000(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_INT32_2147483000);
 
-    int value = vscf_asn1rd_read_int32(asn1rd);
+    int32_t value = vscf_asn1rd_read_int32(asn1rd);
 
     TEST_ASSERT_EQUAL(2147483000, value);
 
@@ -213,7 +213,7 @@ test__asn1rd_read_int32__encoded_int32_neg_2147483000__returns_neg_2147483000(vo
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_INT32_NEG_2147483000);
 
-    int value = vscf_asn1rd_read_int32(asn1rd);
+    int32_t value = vscf_asn1rd_read_int32(asn1rd);
 
     TEST_ASSERT_EQUAL(-2147483000, value);
 
@@ -228,7 +228,7 @@ test__asn1rd_read_uint8__encoded_uint8_255__returns_255(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_UINT8_255);
 
-    int value = vscf_asn1rd_read_uint8(asn1rd);
+    uint8_t value = vscf_asn1rd_read_uint8(asn1rd);
 
     TEST_ASSERT_EQUAL(255, value);
 
@@ -243,7 +243,7 @@ test__asn1rd_read_uint8__encoded_uint8_0__returns_0(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_UINT8_0);
 
-    int value = vscf_asn1rd_read_uint8(asn1rd);
+    uint8_t value = vscf_asn1rd_read_uint8(asn1rd);
 
     TEST_ASSERT_EQUAL(0, value);
 
@@ -258,7 +258,7 @@ test__asn1rd_read_uint16__encoded_uint16_65535__returns_65535(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_UINT16_65535);
 
-    int value = vscf_asn1rd_read_uint16(asn1rd);
+    uint16_t value = vscf_asn1rd_read_uint16(asn1rd);
 
     TEST_ASSERT_EQUAL(65535, value);
 
@@ -273,7 +273,7 @@ test__asn1rd_read_uint32__encoded_uint32_4294967295__returns_4294967295(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_UINT32_4294967295);
 
-    int value = vscf_asn1rd_read_uint32(asn1rd);
+    uint32_t value = vscf_asn1rd_read_uint32(asn1rd);
 
     TEST_ASSERT_EQUAL(4294967295, value);
 
@@ -288,7 +288,7 @@ test__asn1rd_read_uint__encoded_uint_4294967295__returns_4294967295(void) {
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_UINT_4294967295);
 
-    int value = vscf_asn1rd_read_uint(asn1rd);
+    unsigned int value = vscf_asn1rd_read_uint(asn1rd);
 
     TEST_ASSERT_EQUAL(4294967295, value);
 
@@ -303,7 +303,7 @@ test__asn1rd_read_int64__encoded_int64_9223372036854775807__returns_922337203685
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_INT64_9223372036854775807);
 
-    int value = vscf_asn1rd_read_int64(asn1rd);
+    int64_t value = vscf_asn1rd_read_int64(asn1rd);
 
     TEST_ASSERT_EQUAL(9223372036854775807, value);
 
@@ -318,7 +318,7 @@ test__asn1rd_read_int64__encoded_int64_neg_9223372036854775808__returns_neg_9223
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_INT64_NEG_9223372036854775808);
 
-    int value = vscf_asn1rd_read_int64(asn1rd);
+    int64_t value = vscf_asn1rd_read_int64(asn1rd);
 
     TEST_ASSERT_EQUAL(-9223372036854775808, value);
 
@@ -333,7 +333,7 @@ test__asn1rd_read_uint64__encoded_uint64_18446744073709551615__returns_184467440
 
     vscf_asn1rd_reset(asn1rd, test_asn1_encoded_UINT64_18446744073709551615);
 
-    int value = vscf_asn1rd_read_uint64(asn1rd);
+    uint64_t value = vscf_asn1rd_read_uint64(asn1rd);
 
     TEST_ASSERT_EQUAL(18446744073709551615, value);
